requirements.txt
django
mysqlclient
django-bootstrap4
pyyaml
pytest
pytest-django
pytest-mock
ldap3
ansi2html
https://github.com/milliams/python-citc/releases/download/0.3.1/citc-0.3.1-py3-none-any.whl
